requirements.txt
fastapi==0.105.0
sentence-transformers==2.2.2
uvicorn==0.24.0.post1
python-multipart==0.0.6
faiss-cpu==1.7.4
PyPDF2==3.0.1
nltk==3.8.1